requirements.txt
# Secure Edge Vision System - Dependencies
# Install with: pip install -r requirements.txt

# ============ Core ============
opencv-python>=4.8.0
numpy>=1.24.0

# ============ AI/ML ============
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# ============ Web Server ============
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.0
aiofiles>=23.0.0

# ============ Security ============
cryptography>=41.0.0

# ============ Utils ============
python-dotenv>=1.0.0

# ============ Testing ============
pytest>=7.0.0
pytest-asyncio>=0.21.0
